fix processing for decision tables
